% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahon_bedform_relations.r
\name{ancient_1}
\alias{ancient_1}
\title{Estimating sediment flux in ancient rivers.}
\usage{
ancient_1(Tsm, D50, Hbf, uncert = NULL)
}
\arguments{
\item{Tsm}{The mean thickness of cross-bedded cosets. (see publication for more information)}

\item{D50}{The median bedload grain size. Determined from samples.}

\item{Hbf}{The mean bankfull flow depth. Determined from preserved barforms.}

\item{uncert}{The uncertainty window for empirical parameters. The user should put in a confidence interval as a decimal. A lower confidence interval will have a narrower error envelope, but will require more tenuous conclusion.}
}
\value{
Returns a list with elements:
\code{S}, the river slope (dimensionless).
\code{hd}, the dune height in the ancient river (m).
\code{V}, the dune migration rate in the ancient river (m/s).
\code{qs}, the bedload flux in the ancient river (m^3/s).
If uncertainty is also requested, the output will include upper and lower bounds for confidence, \code{V_env} and \code{qs_env}
}
\description{
\code{ancient_1} estimates the sediment flux (\code{qs}) of ancient river deposits based on dune cross-stratification, bankfull depth reconstruction, and grain size. Implemented from Mahon et al. (2018). Original relations referenced therein.
}
