% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahon_bedform_relations.r
\name{bedform_bedload}
\alias{bedform_bedload}
\title{Bedform-bedload equation}
\usage{
bedform_bedload(hd, V, phi = 0.36)
}
\arguments{
\item{hd}{The mean height of bedforms in meters.}

\item{V}{The dune migration rate in meters per second.}

\item{phi}{The porosity of the riverbed material. Default is 0.36.}
}
\value{
Returns a vector \code{qs}, the sediment flux in m^3/s.
}
\description{
\code{bedform_bedload} calculates the sediment flux through bedload in a river based on the height and migration rate of bedforms. From Simons et al. (1965), Implemented from Mahon et al. (2018)
}
