% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahon_bedform_relations.r
\name{bedform_velocity}
\alias{bedform_velocity}
\title{Bedform velocity relation}
\usage{
bedform_velocity(beta0, beta1, S)
}
\arguments{
\item{beta0}{Empirical constant. Value and associated uncertainty from Table 1 in Mahon + McElroy 2018.}

\item{beta1}{Empirical constant. Value and associated uncertainty from Table 1 in Mahon + McElroy 2018.}

\item{S}{The dimensionless river slope.}
}
\value{
Returns a vector \code{V}, the bedform migration rate in m/s.
}
\description{
\code{bedform_velocity} estimates the velocity of a migrating bedform using a power law fit for slope. Implemented from Mahon et al. (2018)
}
