% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahon_bedform_relations.r
\name{dune_height}
\alias{dune_height}
\title{Dune height from cosets relation}
\usage{
dune_height(gamma, Tsm)
}
\arguments{
\item{gamma}{Empirical constant. Value and associated uncertainty from Table 1 in Mahon + McElroy 2018.}

\item{Tsm}{The mean thickness of cross-bedded cosets. (see publication for more information)}
}
\value{
Returns a vector \code{hd}, the dune height in the ancient river.
}
\description{
\code{dune_height} estimates the height of dunes based on cross-stratification. Originally developed by Leclair and Bridge (2001). Implemented from Mahon et al. (2018)
}
