% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleo_depth.r
\name{H2h}
\alias{H2h}
\title{Paleo-depth calculation from dune height}
\usage{
H2h(H, CL = 50, output = "vanilla")
}
\arguments{
\item{H}{A vector of measured cross-set thicknesses in meters.}

\item{CL}{A pre-specified confidence level. If you pick a low confidence (50\%), the range in estimated flow depth will be smaller.}

\item{output}{Takes two values: \code{vanilla} and \code{rich}. \code{vanilla} only returns the estimated value of \code{h}. \code{rich} returns \code{h}, as well as \code{CI}, the confidence interval around the \code{h} estimate.}
}
\value{
Either a single numeric if \code{output = vanilla}, or a list if \code{output = rich}.
}
\description{
\code{H2h} takes estimated dune heights (H) and calculates a flow depth.
}
