% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lynds_slope_method_1.r
\name{lynds_one}
\alias{lynds_one}
\title{Simple method to calculate paleoslope from preserved sediments}
\usage{
lynds_one(D, H, tau_mode = "fixed", rhos = 2650, rhof = 1000)
}
\arguments{
\item{D}{The median grain size of bedload sediments.}

\item{tau_mode}{Denotes whether to use a fixed value for the shields number of the flow, or to use a logarithmically-spaced range going from one to ten.}

\item{rhos}{The density of the sediment in kg/m^3}

\item{rhof}{The density of the fluid in kg/m^3}
}
\value{
Returns a single dimensionless number giving the slope.
}
\description{
\code{lynds_one} calculates the slope of an ancient river, provided you know the flow depth, the grain size of the bed material, and make some assumptions about shear stress in the river. Implemented from Lynds et al. (2014)
}
