% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lynds_slope_method_3.r
\name{lynds_three}
\alias{lynds_three}
\title{Slackwater/suspension method to calculate paleoslope from preserved sediments}
\usage{
lynds_three(Dbed, Dsusp, H, rhos = 2650, rhof = 1000,
  ws_mode = "simple")
}
\arguments{
\item{Dbed}{The median grain size of bedload sediments in meters.}

\item{Dsusp}{The maximum grain size of suspended sediments in meters.}

\item{H}{The paleoflow depth in meters.}

\item{rhos}{The density of the sediment in kg/m^3}

\item{rhof}{The density of the fluid in kg/m^3}

\item{ws_mode}{Allows you to choose one of two different methods for calculating the settling velocity of the bedload particles. 'simple' ignores the Corey Shape Factor and the Powers Index. See documentation for \code{settle} for more information. For now, this feature is disabled.}
}
\value{
Returns a list with two elements: \code{S}, the dimensionless slope, and \code{F}, the ratio of total boundary shear stress to skin friction shear stress.
}
\description{
\code{lynds_three} calculates the slope of an ancient river, provided you know the flow depth, the grain size of the bed material, and the maximum grain size of suspended sediment. Implemented from Lynds et al. (2014)
}
