% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lynds_slope_method_2.r
\name{lynds_two}
\alias{lynds_two}
\title{Variable Shields number method to calculate paleoslope from preserved sediments}
\usage{
lynds_two(D, H, ws_mode = "simple", rhos = 2650, rhof = 1000)
}
\arguments{
\item{D}{The median grain size of bedload sediments in meters.}

\item{H}{The paleoflow depth in meters.}

\item{ws_mode}{Allows you to choose one of two different methods for calculating the settling velocity of the bedload particles. 'simple' ignores the Corey Shape Factor and the Powers Index. See documentation for \code{settle} for more information. For now, this feature is disabled.}

\item{rhos}{The density of the sediment in kg/m^3}

\item{rhof}{The density of the fluid in kg/m^3}
}
\value{
Returns a single dimensionless number giving the slope.
}
\description{
\code{lynds_two} calculates the slope of an ancient river, provided you know the flow depth, the grain size of the bed material, and make some assumptions about shear stress in the river. The important part here is that the function implements a different shields number depending on whether the bed material is fine or coarse. Implemented from Lynds et al. (2014)
}
