% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahon_bedform_relations.r
\name{paleo_slope}
\alias{paleo_slope}
\title{Paleoslope relation}
\usage{
paleo_slope(alpha0, alpha1, alpha2, D50, Hbf)
}
\arguments{
\item{alpha0}{Empirical constant. Value and associated uncertainty from Table 1 in Mahon + McElroy 2018.}

\item{alpha1}{Empirical constant. Value and associated uncertainty from Table 1 in Mahon + McElroy 2018.}

\item{alpha2}{Empirical constant. Value and associated uncertainty from Table 1 in Mahon + McElroy 2018.}

\item{D50}{The median bedload grain size. Determined from samples.}

\item{Hbf}{The mean bankfull flow depth. Determined from preserved barforms.}
}
\value{
Returns a vector \code{S}, the dimensionless paleoslope of the ancient river.
}
\description{
\code{paleo_slope} estimates the slope of an ancient river. Originally developed in Trmapush et al. 2014. Implemented from Mahon et al. (2018)
}
