% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settle.r
\name{settle}
\alias{settle}
\title{Calculate settling velocity for a particle.}
\usage{
settle(ws_mode, D, constants)
}
\arguments{
\item{ws_mode}{Denotes whether to use a 'simpleDietrich' equation, neglecting the shape and roundness of particles, or to use the more 'complexDietrich' formulation. Also allows 'fergusonchurch' for the ferguson and church relation, which is much simpler.}

\item{D}{The grain size you wish to calculate settling velocity for.}

\item{constants}{A list of constants for calculation, need in order (g) gravitational constant, (rhos) sediment density, (rhof) fluid density, (nuu) kinematic viscosity, (csf) corey shape factor, and (PI) Powers roundness index. For the simple formulation, neglect the last two. Defaults to constants = list(g = 9.8, rhos = 2650, rhof = 1000, nuu  = 1.004e-6, csf = 0.7, PI = 3.5)}
}
\value{
Returns a list with two values \code{$ws} and \code{$wstar} which are the settling velocity and dimensionless settling velocity, respectively.
}
\description{
\code{settle} calculates the settling velocity for a particle, given some parameters. Offers a choice between a relation published by Dietrich (1982) or Ferguson and Church (2004)
}
\examples{
settle('simpleDietrich', 120e-6)
settle('complexDietrich', 120e-6, list(g = 9.8, rhos = 2650, rhof = 1000, nuu  = 1.004e-6, csf = 0.7, PI = 3.5)
settle('fergusonchurch', 120e-6)
}
