% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trampush_slope_method.r
\name{trampush_slp}
\alias{trampush_slp}
\title{Global empirical relation method for paleoslope inversion}
\usage{
trampush_slp(Dbed, H, perc = 50)
}
\arguments{
\item{Dbed}{The median grain size of bedload sediments in meters.}

\item{H}{The paleoflow depth in meters.}
}
\value{
Returns a vector \code{S}, the dimensionless slope.
}
\description{
\code{trampush_slp} calculates the slope of an ancient river, provided you know the flow depth and the grain size of the bed material. Implemented from Trampush et al. (2014)
}
