% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleo_depth.r
\name{xset2depthSimple}
\alias{xset2depthSimple}
\title{Paleo-depth calculation from cross-set thicknesses}
\usage{
xset2depthSimple(y)
}
\arguments{
\item{y}{A vector of measured cross-set thicknesses in meters.}
}
\value{
The paleo-depth estimate in meters.
}
\description{
\code{xset2depthSimple} uses the simplest implementation of both \code{xset2H} and \code{H2h}. It uses the vanilla options of each one and spits out an estimate of paleo-depth based on the defaults. Look in the code for the other functions to see what those defaults are.
}
