% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleo_depth.r
\name{xset2H}
\alias{xset2H}
\title{Dune height calculation from cross-set thicknesses}
\usage{
xset2H(y, mode = "simple", output = "vanilla")
}
\arguments{
\item{y}{A vector of measured cross-set thicknesses in meters.}

\item{mode}{Takes two values: \code{simple} and \code{pdfFit}. \code{simple} uses a simple approximation to determine the beta value in Leclair's equation relating x-sets to dune heights. \code{pdfFit} uses a more complicated procedure, fitting a PDF from Paola  + Borgman to get the beta value. \code{pdfFit} is better suited when there are adequate data to estimate the distribution.}

\item{output}{Takes two values: \code{vanilla} and \code{rich}. \code{vanilla} only returns the estimated value of \code{H}, or dune height. \code{rich} returns \code{H}, as well as \code{beta} and either \code{startVal} which is the inital guess of a for the MLE algorithm in \code{pdfFit}, or a span of \code{H} values giving a reasonable range of one particular assumption.}
}
\value{
Either a single numeric if \code{output = vanilla}, or a list if \code{output = rich}.
}
\description{
\code{xset2H} takes measured cross-set thicknesses (\code{y}) and converts it to dune heights (H).
}
